+++ src/orders/models/index.ts
@@ -1,7 +1,8 @@
-export { Orders } from './orders';
-export { Items } from './items';
-export { BillingAddress } from './billing_address';
+export { Order } from './order';
+export { Item } from './item';
+export { Address } from './address';
 export { Payment } from './payment';
-export { ShippingAssignments } from './shipping_assignments';
+export { ShippingAssignment } from './shipping-assignment';
+export { OrderExtensionAttributes } from './order-extension-attributes';
 export { Shipping } from './shipping';
 export { Total } from './total';