+++ src/orders/models/shipping.ts
@@ -1,12 +1,13 @@
 import { Type } from 'class-transformer';
-import { BillingAddress } from './billing_address';
+import { Address } from './address';
 import { Total } from './total';
 
 export class Shipping {
 
-  @Type(() => BillingAddress)
-  address: BillingAddress = new BillingAddress();
   method: string = '';
+
+  @Type(() => Address)
+  address: Address = new Address();
 
   @Type(() => Total)
   total: Total = new Total();