+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-# Changelog
+# Change log
+
+#### 0.8.1
+
+-   Renamed models in products module to be more consistent with the rest of the package.
+
+### 0.8.0 Breaking Changes
+
+-   Renamed models in orders module to be more consistent with the rest of the package.
 
 ### 0.7.0